Added check for existence to delete

+++ src/data/delete.ts
@@ -1,7 +1,14 @@
 import {TrainingSession, NewTrainingSession, DataCallback} from "./types.d"
 import {connection, TABLE_NAME} from './common'
+import * as dbread from './read'
 import * as mysql from 'mysql'
 
 export function deleteTraining(id : number, done: any) {
-    connection.query("DELETE FROM " + TABLE_NAME + " WHERE id=?", [id], done);
+    dbread.exists(id, doesExist=> {
+        if (!doesExist) {
+            done({error: 'Record does not exist'})
+        } else {
+            connection.query("DELETE FROM " + TABLE_NAME + " WHERE id=?", [id], done);
+        }
+    })
 }
